chore: updated for KubeJS 2100.7.0
Note: Stages support has not been re-added. Support in KJS is still WIP.
