--- common/src/main/java/dev/ftb/mods/ftbxmodcompat/ftbfiltersystem/kubejs/CustomFilterEventJS.java
@@ -1,22 +1,0 @@
-package dev.ftb.mods.ftbxmodcompat.ftbfiltersystem.kubejs;
-
-import dev.latvian.mods.kubejs.event.EventJS;
-import net.minecraft.world.item.ItemStack;
-
-public class CustomFilterEventJS extends EventJS {
-    private final ItemStack stack;
-    private final String data;
-
-    public CustomFilterEventJS(ItemStack stack, String data) {
-        this.stack = stack;
-        this.data = data;
-    }
-
-    public ItemStack getStack() {
-        return stack;
-    }
-
-    public String getData() {
-        return data;
-    }
-}
