--- common/src/main/resources/kubejs.classfilter.txt
@@ -1,3 +1,0 @@
-+dev.ftb.mods.ftbquests
-+dev.ftb.mods.ftbchunks
-+dev.ftb.mods.ftbteams
