+++ fabric/src/main/java/dev/ftb/mods/ftbxmodcompat/generic/gamestages/fabric/StagesSetupImpl.java
@@ -1,0 +1,7 @@
+package dev.ftb.mods.ftbxmodcompat.generic.gamestages.fabric;
+
+public class StagesSetupImpl {
+    public static void addKubeJSProvider() {
+        throw new IllegalStateException("KubeJS stages only supported on Neoforge!");
+    }
+}
