+++ gradle.properties
@@ -11,12 +11,12 @@ maven_group=dev.ftb.mods
 curseforge_id=889915
 
 forge_version=50.0.9
-neoforge_version=21.0.14-beta
+neoforge_version=21.0.43-beta
 neoforge_loader_version=4
 fabric_loader_version=0.15.11
 fabric_api_version=0.100.1+1.21
 
-architectury_version=13.0.1
+architectury_version=13.0.2
 
 ftb_library_version=2100.1.1
 ftb_quests_version=2100.1.1
@@ -36,7 +36,7 @@ balm_neoforge_version=5438018
 balm_forge_version=5438000
 balm_fabric_version=5437997
 jei_version=19.0.0.11
+kubejs_version=2100.7.0-build.56
 
 # TODO compiling against 1.20.4 version for now
-kubejs_version=2004.7.0-build.16
 gamestages_version=15.0.1
