+++ neoforge/src/main/java/dev/ftb/mods/ftbxmodcompat/neoforge/ftbquests/kubejs/FTBQuestsKubeJSPlayerData.java
@@ -1,0 +1,22 @@
+package dev.ftb.mods.ftbxmodcompat.neoforge.ftbquests.kubejs;
+
+import dev.ftb.mods.ftbquests.quest.BaseQuestFile;
+import dev.ftb.mods.ftbquests.quest.ServerQuestFile;
+import dev.ftb.mods.ftbquests.quest.TeamData;
+import net.minecraft.world.entity.player.Player;
+
+public class FTBQuestsKubeJSPlayerData extends FTBQuestsKubeJSTeamData {
+	private final Player player;
+
+	public FTBQuestsKubeJSPlayerData(Player p) {
+		player = p;
+	}
+
+	public BaseQuestFile getFile() {
+		return ServerQuestFile.INSTANCE;
+	}
+
+	public TeamData getData() {
+		return ServerQuestFile.INSTANCE.getOrCreateTeamData(player);
+	}
+}
