+++ neoforge/src/main/java/dev/ftb/mods/ftbxmodcompat/neoforge/ftbteams/kubejs/FTBTeamsKubeJSEvents.java
@@ -1,0 +1,11 @@
+package dev.ftb.mods.ftbxmodcompat.neoforge.ftbteams.kubejs;
+
+import dev.latvian.mods.kubejs.event.EventGroup;
+import dev.latvian.mods.kubejs.event.EventHandler;
+
+public interface FTBTeamsKubeJSEvents {
+    EventGroup EVENT_GROUP = EventGroup.of("FTBTeamsEvents");
+
+    EventHandler PLAYER_JOINED_PARTY = EVENT_GROUP.server("playerJoinedParty", () -> PlayerTeamKubeEvent.class);
+    EventHandler PLAYER_LEFT_PARTY = EVENT_GROUP.server("playerLeftParty", () -> PlayerTeamKubeEvent.class);
+}
