+++ neoforge/src/main/resources/kubejs.classfilter.txt
@@ -1,0 +1,4 @@
++dev.ftb.mods.ftbquests
++dev.ftb.mods.ftbchunks
++dev.ftb.mods.ftbteams
++dev.ftb.mods.ftbfiltersystem
